More Updates
Added Permission Checking
Removed Age
Added Delete Commands on startup

+++ modules/commands/destroy.js
@@ -2,11 +2,13 @@ module.exports = {
     destroy : function() {
         client.on('interactionCreate', async (command) => {
             if (!command.isCommand()) return;
-            const { commandName, description, type, options } = command       
-            if(commandName == 'destroy') {
-                command.reply({content:'Shutting down...',ephemeral:true})
-                client.destroy();
-            } 
+            const { commandName, description, type, options } = command
+            if(command.permissions == 'MANAGE_ROLES') {       
+                if(commandName == 'destroy') {
+                    command.reply({content:'Shutting down...',ephemeral:true})
+                    client.destroy();
+                }
+            }
         })
     }   
 }
